Replace Gauge Default doc with imported theme and remove PB styles doc
