--- playbook/app/pb_kits/playbook/pb_gauge/docs/_gauge_pb_styles.md
@@ -1,1 +1,0 @@
-You don't need to use the Gauge Kit to apply Playbook styles to your Highcharts gauge chart. Just import gaugeTheme.ts and merge it with your chart options—Playbook’s styling will apply automatically.
